Added user login, JWT implementation and status

+++ controller/feed.js
@@ -3,6 +3,7 @@ const fs = require("fs");
 const path = require("path");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
     const pageNo = req.query.page || 1;
@@ -17,7 +18,7 @@ exports.getPosts = (req, res, next) => {
                 .limit(postPerPage)
         })
         .then(posts => {
-            res.status(200).json({ message: "Posts found", posts: posts, totalItems:totalPosts })
+            res.status(200).json({ message: "Posts found", posts: posts, totalItems: totalPosts })
         })
         .catch(err => {
             console.log(err);
@@ -40,22 +41,31 @@ exports.createPost = (req, res, next) => {
         err.statusCode = 422;
         throw err;
     }
-    console.log(req.file);
     const imageUrl = req.file.path.replace("\\", "/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Rahul' }
+        creator: req.userId
     });
     post.save()
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
         .then(result => {
             console.log(result);
             res.status(200).json({
                 message: "Post Created",
-                post: result
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -114,6 +124,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearPath(post.imageUrl);
             }
@@ -137,11 +152,64 @@ exports.deletePost = (req, res, next) => {
     const postId = req.params.postId;
     Post.findById(postId)
         .then(post => {
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             clearPath(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({ message: "Post Deleted." });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
+        })
+}
+
+exports.getStatus = (req, res, next) => {
+    User.findById(req.userId)
+        .then(user => {
+            if (!user) {
+                const error = new Error("User not found");
+                error.statusCode = 404;
+                throw error;
+            }
+            res.status(200).json({ status: user.status });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
+        })
+}
+
+exports.updateStatus = (req, res, next) => {
+    const status = req.body.statusData;
+    User.findById(req.userId)
+        .then(user => {
+            if (!user) {
+                const error = new Error("User not found");
+                error.statusCode = 404;
+                throw error;
+            }
+            user.status = status;
+            user.save();
+        })
+        .then(result =>{
+            res.status(200).json({message: "Status updated"});
         })
         .catch(err => {
             if (!err.statusCode) {
